# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Multi-Agent LLM Research Automation Platform API
  description: |
    Comprehensive API documentation for the Multi-Agent Research Platform.
    
    ## Architecture
    
    The platform consists of:
    - **AI Engine** (Python/FastAPI): Core agent orchestration and LLM processing
    - **Backend** (Node.js/Express): User management, research sessions, and data persistence
    - **Frontend** (React/TypeScript): User interface and visualization
    
    ## Authentication
    
    Most endpoints require JWT authentication via the `x-auth-token` header.
    
    ## Features
    
    - Multi-agent research workflows
    - Real-time streaming chat
    - Memory management
    - Web search integration
    - Export functionality (Markdown, PDF, LaTeX)
    - Token usage tracking
  version: "2.0.0"
  contact:
    name: Research Team
    email: research@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: AI Engine (Python/FastAPI)
  - url: http://localhost:5000
    description: Backend (Node.js/Express)
paths:
...
  /auth/login:
    post:
      tags: ["Authentication"]
      summary: User login
      description: Authenticate user and return JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: ["email", "password"]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
            example:
              email: demo@example.com
              password: demo1234
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
              example:
                token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9
                user:
                  id: 1
                  email: demo@example.com
                  name: Demo User
        '400':
          description: Invalid credentials
  /auth/me:
    get:
      tags: ["Authentication"]
      summary: Get current user
      description: Retrieve the authenticated user's profile from the JWT token
      security:
        - JWTAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Not authenticated
        '404':
          description: User not found
...
  /chat/message:
    post:
      tags: ["Chat"]
      summary: Send chat message
      description: Send message and get response (non-streaming)
      security:
        - JWTAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: ["message", "research_id"]
              properties:
                message:
                  type: string
                research_id:
                  type: integer
      responses:
        '200':
          description: Chat response
          content:
            application/json:
              schema:
                type: object
                properties:
                  response:
                    type: string
                  message_id:
                    type: integer
  /chat/stream:
    post:
      tags: ["Chat"]
      summary: Streaming chat
      description: Send message and stream response in real-time
      security:
        - JWTAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: ["message", "research_id"]
              properties:
                message:
                  type: string
                research_id:
                  type: integer
      responses:
        '200':
          description: Streaming response
          content:
            text/event-stream:
              schema:
                type: string
  
  # Memory
...
components:
  securitySchemes:
    JWTAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token in x-auth-token header
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
          format: email
        name:
          type: string
        created_at:
          type: string
          format: date-time
...